backend/audiobook_project/requirements.txt
absl-py==0.13.0
aiohttp==3.9.3
aiosignal==1.3.1
alpaca-trade-api==3.2.0
aniso8601==9.0.1
argon2-cffi==20.1.0
asgiref==3.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==3.3.0
bs4==0.0.1
CacheControl==0.12.11
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==37.0.2
cycler==0.10.0
db-dtypes==1.1.1
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
dj-database-url==2.2.0
Django==3.2.4
django-cors-headers==4.4.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.18
filelock==3.4.0
firebase-admin==6.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flatbuffers==1.12
frozenlist==1.4.1
future==1.0.0
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.9.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-oauth2l==1.0.2
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.1.0
httplib2==0.22.0
huggingface-hub==0.2.1
idna==2.10
igraph==0.10.4
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==5.2.2
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.3
multidict==6.0.5
mysqlclient==2.2.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.3.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.9
pandas==1.2.4
pandas-gbq==0.19.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.2
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.22.2
protobuf==4.22.1
PuLP==2.8.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.7.0
Pygments==2.9.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-pptx==0.6.23
pytz==2021.1
pywin32==300
pywinpty==1.0.1
PyYAML==6.0.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
razorpay==1.3.0
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==4.1.5
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textblob==0.17.1
texttable==1.6.7
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.13.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.12.2
upstox==2.0.3
upstox-python-sdk==2.4.1
uritemplate==4.1.1
urllib3==1.26.6
uuid==1.30
virtualenv==20.14.1
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==0.16.1
whitenoise==6.7.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.1.0
xgboost==1.4.2
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
